fix: get 요청시 토큰이 헤더에 한박자 늦게 탑재되는 에러 해결
