+++ src/api/axios.ts
@@ -1,9 +1,15 @@
 import axios from "axios";
 
-const token = "test";
 const client = axios.create({
   baseURL: "https://sp-taskify-api.vercel.app/4-7/",
-  headers: { Authorization: "Bearer " + token },
+});
+
+client.interceptors.request.use(async (config) => {
+  const token = localStorage.getItem("token");
+  if (config.headers) {
+    config.headers.Authorization = `Bearer ${token}`;
+  }
+  return config;
 });
 
 export default client;
